pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "google-generativeai>=0.8.6",
    "pytelegrambotapi>=4.32.0",
]
